// src/App/Body/Interface/SmallPreview/interfaces.tsx
export interface IContainer {
    display: any;
    widthCanvasPixel: number;
    heightCanvasPixel: number;
}
export interface IFragment {
    widthCanvasPixel?: number;
    heightCanvasPixel?: number;
    // onMouseDown: (e: React.MouseEvent) => void;
    // onMouseUp: (e: React.MouseEvent) => void;
    // onMouseLeave: (e: React.MouseEvent) => void;
    // onMouseMove: (e: React.MouseEvent) => void;
}